perf: 移除 mysql、mariadb、postgresql、sqlserver 的支持
