+++ pkg/proxy/server.go
@@ -358,8 +358,7 @@ func (s *Server) checkRequiredAuth() error {
 			utils.IgnoreErrWriteString(s.UserConn, msg)
 			return errors.New("no auth token")
 		}
-	case srvconn.ProtocolMySQL, srvconn.ProtocolMariadb, srvconn.ProtocolTELNET,
-		srvconn.ProtocolSQLServer, srvconn.ProtocolPostgreSQL, srvconn.ProtocolClickHouse,
+	case srvconn.ProtocolTELNET, srvconn.ProtocolClickHouse,
 		srvconn.ProtocolMongoDB:
 		if err := s.getUsernameIfNeed(); err != nil {
 			msg := utils.WrapperWarn(lang.T("Get auth username failed"))
@@ -543,43 +542,6 @@ func (s *Server) getContainerConn(clusterServer string) (
 	return
 }
 
-func (s *Server) getMySQLConn(localTunnelAddr *net.TCPAddr) (srvConn *srvconn.MySQLConn, err error) {
-	asset := s.connOpts.authInfo.Asset
-	protocol := s.connOpts.authInfo.Protocol
-	host := asset.Address
-	port := asset.ProtocolPort(protocol)
-	if localTunnelAddr != nil {
-		host = "127.0.0.1"
-		port = localTunnelAddr.Port
-	}
-	mysqlOpts := make([]srvconn.SqlOption, 0, 7)
-	mysqlOpts = append(mysqlOpts, srvconn.SqlHost(host))
-	mysqlOpts = append(mysqlOpts, srvconn.SqlPort(port))
-	mysqlOpts = append(mysqlOpts, srvconn.SqlUsername(s.account.Username))
-	mysqlOpts = append(mysqlOpts, srvconn.SqlPassword(s.account.Secret))
-	mysqlOpts = append(mysqlOpts, srvconn.SqlDBName(asset.SpecInfo.DBName))
-	mysqlOpts = append(mysqlOpts, srvconn.SqlPtyWin(srvconn.Windows{
-		Width:  s.UserConn.Pty().Window.Width,
-		Height: s.UserConn.Pty().Window.Height,
-	}))
-	tokenConnOpts := s.connOpts.authInfo.ConnectOptions
-	switch {
-	case tokenConnOpts.DisableAutoHash != nil:
-		if *tokenConnOpts.DisableAutoHash {
-			mysqlOpts = append(mysqlOpts, srvconn.MySQLDisableAutoReHash())
-		}
-		logger.Debugf("Connection token set disableAutoHash: %v", *tokenConnOpts.DisableAutoHash)
-	case s.connOpts.params != nil:
-		if s.connOpts.params.DisableMySQLAutoHash {
-			mysqlOpts = append(mysqlOpts, srvconn.MySQLDisableAutoReHash())
-			logger.Debugf("Connection params set disableAutoHash: true")
-		}
-
-	}
-	srvConn, err = srvconn.NewMySQLConnection(mysqlOpts...)
-	return
-}
-
 func (s *Server) getRedisConn(localTunnelAddr *net.TCPAddr) (srvConn *srvconn.RedisConn, err error) {
 	asset := s.connOpts.authInfo.Asset
 	protocol := s.connOpts.authInfo.Protocol
@@ -633,52 +595,6 @@ func (s *Server) getMongoDBConn(localTunnelAddr *net.TCPAddr) (srvConn *srvconn.
 		srvconn.SqlCaCert(asset.SecretInfo.CaCert),
 		srvconn.SqlCertKey(asset.SecretInfo.ClientKey),
 		srvconn.SqlAllowInvalidCert(asset.SpecInfo.AllowInvalidCert),
-		srvconn.SqlPtyWin(srvconn.Windows{
-			Width:  s.UserConn.Pty().Window.Width,
-			Height: s.UserConn.Pty().Window.Height,
-		}),
-	)
-	return
-}
-
-func (s *Server) getSQLServerConn(localTunnelAddr *net.TCPAddr) (srvConn *srvconn.SQLServerConn, err error) {
-	asset := s.connOpts.authInfo.Asset
-	protocol := s.connOpts.authInfo.Protocol
-	host := asset.Address
-	port := asset.ProtocolPort(protocol)
-	if localTunnelAddr != nil {
-		host = "127.0.0.1"
-		port = localTunnelAddr.Port
-	}
-	srvConn, err = srvconn.NewSQLServerConnection(
-		srvconn.SqlHost(host),
-		srvconn.SqlPort(port),
-		srvconn.SqlUsername(s.account.Username),
-		srvconn.SqlPassword(s.account.Secret),
-		srvconn.SqlDBName(asset.SpecInfo.DBName),
-		srvconn.SqlPtyWin(srvconn.Windows{
-			Width:  s.UserConn.Pty().Window.Width,
-			Height: s.UserConn.Pty().Window.Height,
-		}),
-	)
-	return
-}
-
-func (s *Server) getPostgreSQLConn(localTunnelAddr *net.TCPAddr) (srvConn *srvconn.PostgreSQLConn, err error) {
-	asset := s.connOpts.authInfo.Asset
-	protocol := s.connOpts.authInfo.Protocol
-	host := asset.Address
-	port := asset.ProtocolPort(protocol)
-	if localTunnelAddr != nil {
-		host = "127.0.0.1"
-		port = localTunnelAddr.Port
-	}
-	srvConn, err = srvconn.NewPostgreSQLConnection(
-		srvconn.SqlHost(host),
-		srvconn.SqlPort(port),
-		srvconn.SqlUsername(s.account.Username),
-		srvconn.SqlPassword(s.account.Secret),
-		srvconn.SqlDBName(asset.SpecInfo.DBName),
 		srvconn.SqlPtyWin(srvconn.Windows{
 			Width:  s.UserConn.Pty().Window.Width,
 			Height: s.UserConn.Pty().Window.Height,
@@ -962,16 +878,10 @@ func (s *Server) getServerConn(proxyAddr *net.TCPAddr) (srvconn.ServerConnection
 		return s.getTelnetConn()
 	case srvconn.ProtocolK8s:
 		return s.getK8sConConn(proxyAddr)
-	case srvconn.ProtocolMySQL, srvconn.ProtocolMariadb:
-		return s.getMySQLConn(proxyAddr)
-	case srvconn.ProtocolSQLServer:
-		return s.getSQLServerConn(proxyAddr)
 	case srvconn.ProtocolRedis:
 		return s.getRedisConn(proxyAddr)
 	case srvconn.ProtocolMongoDB:
 		return s.getMongoDBConn(proxyAddr)
-	case srvconn.ProtocolPostgreSQL:
-		return s.getPostgreSQLConn(proxyAddr)
 	case srvconn.ProtocolClickHouse:
 		return s.getClickHouseConn(proxyAddr)
 	default:
